+++ src/model/Snackbar/Snackbar.tsx
@@ -1,6 +1,6 @@
 import { Button, IconButton } from '@mui/material';
 import { makeAutoObservable } from 'mobx';
-import { SnackbarAction, VariantType, enqueueSnackbar, closeSnackbar } from 'notistack';
+import { SnackbarAction, VariantType, enqueueSnackbar, closeSnackbar, SnackbarMessage } from 'notistack';
 import CloseIcon from '@mui/icons-material/Close';
 
 export default class Snackbar {
@@ -19,11 +19,11 @@ export default class Snackbar {
   /**
    * Enqueues a message to the snackbar used for temporary status updates to the user.
    *
-   * @param msg The message you want to display. Use "\n" to insert new lines.
+   * @param msg The message you want to display. Use "\n" to insert new lines. Can either be a string or a JSX element.
    * @param variant The variant (e.g. error, warning) of snackbar you want to display.
    * @param action The action to display on the snackbar (see notistack docs). If undefined, a close button will be displayed.
    */
-  sendToSnackbar(msg: string, variant: VariantType, action?: SnackbarAction, persist?: boolean) {
+  sendToSnackbar(msg: SnackbarMessage, variant: VariantType, action?: SnackbarAction, persist?: boolean) {
     if (!(enqueueSnackbar instanceof Function)) {
       console.error('enqueueSnackbar is not a function. Cannot send message: ' + msg)
       return;