+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [4.17.1] - 2024-06-11
+
+### Changed
+
+- Added more info about browser compatibility to the error message if navigator.serial is not available. Also added this same info to the homepage.
+
 ## [4.17.0] - 2024-05-31
 
 ### Added
@@ -714,7 +720,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added auto-scroll to TX pane, closes #89.
 - Added special delete behaviour for backspace button when in "send on enter" mode, closes #90.
 
-[unreleased]: https://github.com/gbmhunter/NinjaTerm/compare/v4.17.0...HEAD
+[unreleased]: https://github.com/gbmhunter/NinjaTerm/compare/v4.17.1...HEAD
+[4.17.1]: https://github.com/gbmhunter/NinjaTerm/compare/v4.17.0...v4.17.1
 [4.17.0]: https://github.com/gbmhunter/NinjaTerm/compare/v4.16.0...v4.17.0
 [4.16.0]: https://github.com/gbmhunter/NinjaTerm/compare/v4.15.0...v4.16.0
 [4.15.0]: https://github.com/gbmhunter/NinjaTerm/compare/v4.14.0...v4.15.0